[FastTokenizer]Fix json of tokenizers (#4165)
* Fix json of tokenizers

* Add c_wrap

* Update version
